Made send buffers intent(in)
Prior to this they had been intent(inout) but this was causing problems
with one of our calls in oops.  It makes sense - the send buffers should
not be changed by the send operation.  Still, to minimize impact, I only
changed it for send_real32_r1 and send_real64_r1 (for now)